Finish Day 14
I did not expect any days to call back on / require code from a part 2 of a previous day!
I attempted to refactor my knot hashing for Day 10 to "optimize" it, but the amount of checks that Kotlin puts into the bytecode makes performance comparable to Java tough.
I don't have a Java solution to compare to and don't care to write one.